reset dataLength

+++ src/ENet.NET/ENetBuffer.cs
@@ -5,6 +5,6 @@ namespace ENet.NET
     public struct ENetBuffer
     {
         public ArraySegment<byte> data;
-        public long dataLength;
+        public int dataLength;
     }
 }
