+++ src/ENet.NET/ENets.cs
@@ -24,10 +24,10 @@ namespace ENet.NET
     public delegate void PacketDestroyDelegate(ENetPacket packet);
 
     /** Compresses from inBuffers[0:inBufferCount-1], containing inLimit bytes, to outData, outputting at most outLimit bytes. Should return 0 on failure. */
-    public delegate long CompressorCompressDelegate(object context, ref ENetBuffer inBuffers, long inBufferCount, long inLimit, ArraySegment<byte> outData, long outLimit);
+    public delegate int CompressorCompressDelegate(object context, ref ENetBuffer inBuffers, long inBufferCount, long inLimit, ArraySegment<byte> outData, long outLimit);
 
     /** Decompresses from inData, containing inLimit bytes, to outData, outputting at most outLimit bytes. Should return 0 on failure. */
-    public delegate long CompressorDecompressDelegate(object context, ArraySegment<byte> inData, long inLimit, ArraySegment<byte> outData, long outLimit);
+    public delegate int CompressorDecompressDelegate(object context, ArraySegment<byte> inData, long inLimit, ArraySegment<byte> outData, long outLimit);
 
     /** Destroys the context when compression is disabled or the host is destroyed. May be NULL. */
     public delegate void CompressorDestroyDelegate(object context);
